gimnasio_back/requirements.txt
asgiref==3.8.1
boto3==1.34.143
botocore==1.34.143
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2.3
django-cors-headers==4.4.0
django-simple-history==3.7.0
django-storages==1.14.4
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
drf-yasg==1.21.7
exceptiongroup==1.2.1
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.2
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
Pillow==10.0.0  # Añadir Pillow aquí
